Fix int to isize

+++ benches/common/macros.rs
@@ -17,7 +17,7 @@ macro_rules! bench_binop(
     ($name: ident, $t1: ty, $t2: ty, $binop: ident) => {
         #[bench]
         fn $name(bh: &mut Bencher) {
-            const LEN: uint = 1 << 13;
+            const LEN: usize = 1 << 13;
 
             let mut rng = IsaacRng::new_unseeded();
 
@@ -40,7 +40,7 @@ macro_rules! bench_binop_deref(
     ($name: ident, $t1: ty, $t2: ty, $binop: ident) => {
         #[bench]
         fn $name(bh: &mut Bencher) {
-            const LEN: uint = 1 << 13;
+            const LEN: usize = 1 << 13;
 
             let mut rng = IsaacRng::new_unseeded();
 
@@ -63,7 +63,7 @@ macro_rules! bench_unop(
     ($name: ident, $t: ty, $unop: ident) => {
         #[bench]
         fn $name(bh: &mut Bencher) {
-            const LEN: uint = 1 << 13;
+            const LEN: usize = 1 << 13;
 
             let mut rng = IsaacRng::new_unseeded();
 
@@ -85,7 +85,7 @@ macro_rules! bench_construction(
     ($name: ident, $t: ty, $constructor: path [ $($args: ident: $types: ty),+ ]) => {
         #[bench]
         fn $name(bh: &mut Bencher) {
-            const LEN: uint = 1 << 13;
+            const LEN: usize = 1 << 13;
 
             let mut rng = IsaacRng::new_unseeded();
